+++ src/device_drama/plugins/01_main_hardware/02_cpu_info.py
@@ -11,8 +11,8 @@ class DDPlugin(BasePlugin):
         self.info.author = 'Tadeusz Miszczyk'
         self.info.description = 'Return CPU info'
         self.info.name = 'CPU-Info'
-        self.info.plugin_version = '23.9.4'
-        self.info.compatible_version = '23.9.4'
+        self.info.plugin_version = '23.9.12'
+        self.info.compatible_version = '23.9.12'
 
     def run(self) -> List[Dict[str, str]]:
         with open('/proc/cpuinfo', 'r') as file: